Allow PUT requests to be handled by a CGI
If a CGI has been assigned to handle a given URI, let it handle
PUT requests as well as GET / POST. This allows a cgi to act as
a virtual directory, rather than limiting PUT requests to store
data as files in a normal directory.

The http 1.1 standard does not address whether virtual directories
served by a cgi program should be supported, but at least it doesn't
seem to forbid it, only forbidding automatic redirection from one
directory to another. Thus rewrite rules should be disabled for PUT,
but sending the request to the client-specified path that happens
to be a virtual directory served by an app and stored in a database
(rather than a normal directory in a mounted filesystem) seems legitimate.

Since virtual PUTs seem allowable it is desirable to implement this
feature, due to PUT being a fairly common part of the popular REST app
development strategy. For a specific example of such an app, serving
PUT requests in the same manner as GET and POST requests is required
to be able to serve the ReviewBoard application https://www.reviewboard.org/
which stores uploaded review requests in its database rather than as
external files.